feat: move serialization to coin-framework
