+++ libs/coin-bitcoin/src/descriptor.test.ts
@@ -3,31 +3,34 @@ import {
   getCryptoCurrencyById,
   setSupportedCurrencies,
 } from "@ledgerhq/coin-framework/currencies/index";
+import { fromAccountRaw } from "@ledgerhq/coin-framework/serialization/index";
 import { inferDescriptorFromDeviceInfo, inferDescriptorFromAccount } from "./descriptor";
-// import { fromAccountRaw } from "../../account";
 import bitcoinDatasets from "./datasets/bitcoin";
+import { assignFromAccountRaw } from "./serialization";
 
 setSupportedCurrencies(["bitcoin"]);
-xdescribe("inferDescriptorFromAccount", () => {
+describe("inferDescriptorFromAccount", () => {
   invariant(bitcoinDatasets.accounts, "bitcoin datasets have accounts");
   // FIXME: migrate away from invariant for more type guard
-  // const [bitcoin1, bitcoin2] = bitcoinDatasets.accounts.map(a => fromAccountRaw(a.raw));
-  // it("should work on bitcoin account 1", () => {
-  //   expect(inferDescriptorFromAccount(bitcoin1)).toEqual({
-  //     external:
-  //       "pkh([224b6226/44'/0'/0']xpub6BuPWhjLqutPV8SF4RMrrn8c3t7uBZbz4CBbThpbg9GYjqRMncra9mjgSfWSK7uMDz37hhzJ8wvkbDDQQJt6VgwLoszvmPiSBtLA1bPLLSn/0/*)",
-  //     internal:
-  //       "pkh([224b6226/44'/0'/0']xpub6BuPWhjLqutPV8SF4RMrrn8c3t7uBZbz4CBbThpbg9GYjqRMncra9mjgSfWSK7uMDz37hhzJ8wvkbDDQQJt6VgwLoszvmPiSBtLA1bPLLSn/1/*)",
-  //   });
-  // });
-  // it("should work on bitcoin account 2", () => {
-  //   expect(inferDescriptorFromAccount(bitcoin2)).toEqual({
-  //     external:
-  //       "wpkh([d6a9e45e/84'/0'/1']xpub6DEHKg8fgKcb9at2u9Xhjtx4tXGyWqUPQAx2zNCzr41gQRyCqpCn7onSoJU4VS96GXyCtAhhFxErnG2pGVvVexaqF7DEfqGGnGk7Havn7C2/0/*)",
-  //     internal:
-  //       "wpkh([d6a9e45e/84'/0'/1']xpub6DEHKg8fgKcb9at2u9Xhjtx4tXGyWqUPQAx2zNCzr41gQRyCqpCn7onSoJU4VS96GXyCtAhhFxErnG2pGVvVexaqF7DEfqGGnGk7Havn7C2/1/*)",
-  //   });
-  // });
+  const [bitcoin1, bitcoin2] = bitcoinDatasets.accounts.map(a =>
+    fromAccountRaw(a.raw, { assignFromAccountRaw }),
+  );
+  it("should work on bitcoin account 1", () => {
+    expect(inferDescriptorFromAccount(bitcoin1)).toEqual({
+      external:
+        "pkh([224b6226/44'/0'/0']xpub6BuPWhjLqutPV8SF4RMrrn8c3t7uBZbz4CBbThpbg9GYjqRMncra9mjgSfWSK7uMDz37hhzJ8wvkbDDQQJt6VgwLoszvmPiSBtLA1bPLLSn/0/*)",
+      internal:
+        "pkh([224b6226/44'/0'/0']xpub6BuPWhjLqutPV8SF4RMrrn8c3t7uBZbz4CBbThpbg9GYjqRMncra9mjgSfWSK7uMDz37hhzJ8wvkbDDQQJt6VgwLoszvmPiSBtLA1bPLLSn/1/*)",
+    });
+  });
+  it("should work on bitcoin account 2", () => {
+    expect(inferDescriptorFromAccount(bitcoin2)).toEqual({
+      external:
+        "wpkh([d6a9e45e/84'/0'/1']xpub6DEHKg8fgKcb9at2u9Xhjtx4tXGyWqUPQAx2zNCzr41gQRyCqpCn7onSoJU4VS96GXyCtAhhFxErnG2pGVvVexaqF7DEfqGGnGk7Havn7C2/0/*)",
+      internal:
+        "wpkh([d6a9e45e/84'/0'/1']xpub6DEHKg8fgKcb9at2u9Xhjtx4tXGyWqUPQAx2zNCzr41gQRyCqpCn7onSoJU4VS96GXyCtAhhFxErnG2pGVvVexaqF7DEfqGGnGk7Havn7C2/1/*)",
+    });
+  });
 });
 describe("inferDescriptorFromDeviceInfo", () => {
   it("should work on a first Bitcoin legacy derivation", () => {
