+++ libs/coin-framework/src/account/index.ts
@@ -5,5 +5,4 @@ export * from "./groupOperations";
 export * from "./helpers";
 export * from "./ordering";
 export * from "./pending";
-export * from "./serialization";
 export * from "./support";
