+++ libs/ledger-live-common/src/exchange/swap/types.ts
@@ -233,26 +233,6 @@ export type MappedSwapOperation = {
   fromAmount: BigNumber;
   toAmount: BigNumber;
 };
-export type SwapOperation = {
-  provider: string;
-  swapId: string;
-  status: string;
-  receiverAccountId: string;
-  tokenId?: string;
-  operationId: string;
-  fromAmount: BigNumber;
-  toAmount: BigNumber;
-};
-export type SwapOperationRaw = {
-  provider: string;
-  swapId: string;
-  status: string;
-  receiverAccountId: string;
-  tokenId?: string;
-  operationId: string;
-  fromAmount: string;
-  toAmount: string;
-};
 export type SwapState = {
   // NB fromAccount and fromParentAccount and amount come from `useBridgeTransaction`
   useAllAmount?: boolean;
