+++ libs/ledgerjs/packages/types-live/src/index.ts
@@ -14,5 +14,5 @@ export * from "./postOnboarding";
 export * from "./storyly";
 export * from "./chainwatch";
 export * from "./messages";
-export * from "./storyly";
+export * from "./swap";
 export * from "./ABTesting";
